requirements.txt
asgiref==3.7.2
attrs==23.1.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-filter==23.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.7
fcm-django==1.0.12
firebase-admin==5.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Markdown==3.4.4
msgpack==1.0.5
packaging==23.1
phonenumbers==8.13.18
proto-plus==1.22.3
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.5.0
pyparsing==3.1.1
python-dotenv==0.21.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
rsa==4.9
secure-smtplib==0.1.1
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
